preparing for merge

+++ mechasuite/kmc.h
@@ -44,7 +44,7 @@ public:
     std::vector<std::vector<int>> history;
     std::vector<double> times;
 
-    short saveFreq = 1;
+    size_t saveFreq = 1;
     size_t step = 0;
 
     System() = default;
